requirements.txt
# Core Flask and extensions
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1

# Database and migrations
SQLAlchemy==2.0.28
alembic==1.13.1

# Security and validation
Werkzeug==3.0.1
cryptography==42.0.5
email-validator==2.1.0.post1
itsdangerous==2.1.2

# Environment and utilities
python-dotenv==1.0.1
python-dateutil==2.8.2
blinker==1.7.0

# Template engine
Jinja2==3.1.3
MarkupSafe==2.1.5

# CLI
click==8.1.7

wtforms~=3.2.1